🗺️ Atlas: Encapsulate internal submodules and fix AssemblyError dependency cycle
🕸️ Tangle: `pub mod` was overused across the compiler (e.g., `tools`, `errors`, `semantic`, `morphology`), unnecessarily exposing internal implementations. In addition, `AssemblyError` was nested inside `errors.rs` but tightly coupled with `semantic::assembly`, creating a confusing module boundary.

📐 Blueprint:
1. Switched internal submodules across the codebase to use `pub(crate) mod` by default.
2. Restored `pub mod` only for `parser`, `semantic`, `codegen`, and `morphology` to maintain the public API consumed by tests and CLI tools.
3. Extracted `AssemblyError` to its own file (`src/errors/assembly.rs`) and modularized `errors` as a directory facade (`src/errors/mod.rs`), cleanly breaking the module blob.

🧱 Stability: Enforces strict module boundaries, prevents API leaks, and flattens the `errors` module dependency structure.

🔬 Verification: Ran `cargo check`, `cargo clippy`, and `cargo test` successfully.
